+++ Server/components/VectorComponent.h
@@ -3,8 +3,8 @@
 class VectorComponent final
 {
 public:
-    float x = 0;
-    float y = 0;
+    double x = 0;
+    double y = 0;
 };
 class VectorChange final
 {
